+++ ntex-net/src/iocp/io.rs
@@ -58,7 +58,7 @@ fn poll_readiness(ctl: &StreamCtl, ctx: &IoContext, cx: &mut Context<'_>) -> Pol
             ctl.read();
             Poll::Pending
         }
-        Poll::Ready(Readiness::Shutdown | Readiness::Terminate) => Poll::Ready(()),
+        Poll::Ready(Readiness::Close) => Poll::Ready(()),
         Poll::Pending => {
             ctl.pause();
             Poll::Pending
@@ -70,7 +70,7 @@ fn poll_readiness(ctl: &StreamCtl, ctx: &IoContext, cx: &mut Context<'_>) -> Pol
             ctl.write();
             Poll::Pending
         }
-        Poll::Ready(Readiness::Shutdown | Readiness::Terminate) => Poll::Ready(()),
+        Poll::Ready(Readiness::Close) => Poll::Ready(()),
         Poll::Pending => Poll::Pending,
     };