+++ ntex-net/src/polling/io.rs
@@ -56,7 +56,7 @@ async fn run(ctl: StreamCtl, context: ntex_io::IoContext) {
                 readable = true;
                 Poll::Pending
             }
-            Poll::Ready(Readiness::Shutdown | Readiness::Terminate) => Poll::Ready(()),
+            Poll::Ready(Readiness::Close) => Poll::Ready(()),
             Poll::Pending => {
                 modify = true;
                 Poll::Pending
@@ -69,7 +69,7 @@ async fn run(ctl: StreamCtl, context: ntex_io::IoContext) {
                 writable = true;
                 Poll::Pending
             }
-            Poll::Ready(Readiness::Shutdown | Readiness::Terminate) => Poll::Ready(()),
+            Poll::Ready(Readiness::Close) => Poll::Ready(()),
             Poll::Pending => {
                 modify = true;
                 Poll::Pending